requirements.txt
pyspark>=3.4.0
requests>=2.31.0
python-dotenv>=1.0.0
folium>=0.15.0
seaborn>=0.13.0
matplotlib>=3.8.0
urllib3>=2.0.0